% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_struct_weights.R
\name{calc_struct_weights}
\alias{calc_struct_weights}
\title{PLS calculation of weights}
\usage{
calc_struct_weights(X, Ys, QQ, HH, aH, sumabsw = sqrt(dim(X)[2]),
  niter = 20)
}
\arguments{
\item{X}{a (preprocessed) n by px dataset}

\item{Ys}{a (preprocessed) n by py dataset of responses}

\item{QQ}{a py by py symmetric feature kernel representing similarities between features.}

\item{HH}{a px by px symmetric sample kernel representing similarities between samples.}

\item{sumabsw}{a scalar contraint on the L1 norm of the weights.}

\item{niter}{= 20 (default) number of iterations}
}
\description{
Inner engine for PLS calculating weights based on a combined objective, maximising 1) covariance between X and Y, 2) description of feature to feature relationship, 3) Alternative kernel representation of sample-data and 4) with possible sparseness.
}
