% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublecentering.R
\name{doublecentering}
\alias{doublecentering}
\title{Double centring of symmetric distance/similarity matrices}
\usage{
doublecentering(M)
}
\arguments{
\item{M}{a n by n symmetric matrix}
}
\description{
This functions centers a n by n symmetric matrix X resulting that coloumn and row-means are zero. #'
}
