% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innerCCA.R
\name{innerCCA}
\alias{innerCCA}
\title{Sparse Canonical Covariates Analysis (CCA)}
\usage{
innerCCA(CxyR, X, Y, ncomp, sumabs)
}
\arguments{
\item{X}{a n by px matrix}

\item{Y}{a n by py matrix}

\item{ncomp}{scalar - number of components}

\item{sumabs}{1 by 2 vector of L1 contraint on u and v respectively}

\item{Cxy}{a px by py crossproduct matrix between features in X and Y}
}
\description{
This functions minimises |Cxy-udv'|2 s.t. sum(abs(u))<sumabs[1], sum(abs(v))<sumabs[2], norm(u,2)=norm(v,2) = 1 with respect to u and v. sequencially by deflation and subsequently calculates canonical variates.
}
