% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCCA.R
\name{pCCA}
\alias{pCCA}
\title{Regularaized Sparse Canonical Covariates Analysis}
\usage{
pCCA(X, Y, R = diag(1, ncol(Y)), ncomp = 1, sumabs = c(sqrt(ncol(X)),
  sqrt(ncol(Y))), na = 30, nperm = 100)
}
\arguments{
\item{X}{a n by px matrix}

\item{Y}{a n by py matrix}

\item{R}{a py by py symmetric matrix indicating strucural relationship between features in Y.}

\item{ncomp}{= 1 (default) - number of components}

\item{sumabs}{= sqrt(c(px,py)) - vector of L1 contraint on the canonical loadings u and v respectively}

\item{na}{= 30 (default) - indicating the grid of structural penalties based on R to tests or a vector of specific penalties.}

\item{nperm}{= 100 (default) - number of permutations to be perfomed}
}
\description{
This functions calculates a CCA model with L1 norm sparse penalties on the canonical loadings, as well as L2 norm penalties on the structural loadings in Y.
}
\examples{

}
