% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_nipals.R
\name{pen_nipals}
\alias{pen_nipals}
\title{NIPALS algorithm for sparse bilinear one component model of a matrix}
\usage{
pen_nipals(x, sumabs = sqrt(c(nrow(x), ncol(x))), niter = 100)
}
\arguments{
\item{x}{a n by p dataset}

\item{sumabs}{1 by 2 vector ofL1 contraint on u and v respectively}

\item{niter}{= 100 (default) number of iterations}
}
\description{
This functions minimises |x-udv'|2 s.t. sum(abs(u))<sumabs[1], sum(abs(v))<sumabs[2], norm(u,2)=norm(v,2) = 1 with respect to u and v.
}
