% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permCCA.R
\name{permCCA}
\alias{permCCA}
\title{Permutation Test for Regularaized Sparse Canonical Covariates Analysis}
\usage{
permCCA(X, Y, R, ncomp = 1, nperm = 100, sumabs)
}
\arguments{
\item{X}{a n by px matrix}

\item{Y}{a n by py matrix}

\item{R}{a py by py symmetric matrix indicating strucural relationship between features in Y.}

\item{ncomp}{scalar - number of components}

\item{nperm}{scalar = 100 (default) number of permutations to be perfomed}

\item{sumabs}{1 by 2 vector of L1 contraint on the canonical loadings u and v respectively}
}
\description{
This functions calculates a CCA model with L1 norm sparse penalties on the canonical loadings, as well as L2 norm penalties on the structural loadings in Y.
}
\examples{

}
