% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchL.R
\name{searchL}
\alias{searchL}
\title{Find parameter for Soft thressholding}
\usage{
searchL(u, sumabs, niter = 100)
}
\arguments{
\item{u}{a vector or matrix and}

\item{sumabs}{the required contraint}

\item{niter=100}{(default) number of iterations}
}
\description{
This functions searches penalty to match sum(abs(u)) contraint via a griddy split half algorithm
}
