% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchLbic.R
\name{searchLbic}
\alias{searchLbic}
\title{Find parameter for Soft thressholding using Bayes Information Criterion}
\usage{
searchLbic(u, x, v, m = 30, n = dim(x)[1])
}
\arguments{
\item{n}{number of observations}

\item{w}{an initial guess}

\item{M}{a cross product matrix}

\item{m=30}{(default) number of splits to tests}
}
\description{
This functions searches the optimal penalty based on BIC
}
