% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr.R
\name{smr}
\alias{smr}
\title{Sparse Matrix Regression}
\usage{
smr(x, v, sumabs)
}
\arguments{
\item{x}{a n by p dataset}

\item{v}{a 1 by p vector}

\item{a}{scalar value for the L1 contraint on u}
}
\description{
This functions minimises |x-uv'|2 s.t. sum(abs(u))<sumabs.
}
