% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softth.R
\name{softth}
\alias{softth}
\title{Soft thressholding}
\usage{
softth(x, d)
}
\arguments{
\item{x}{a vector or matrix and d a scalar}
}
\description{
This functions truncates values in x (numerically) less than d to zero and reduces the magnitude for values larger than d by d.
}
