% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structpls.R
\name{structpls}
\alias{structpls}
\title{Structural PLS}
\usage{
structpls(X, Y, lv, Q, H, aH, sumabsw, deflateX = FALSE)
}
\arguments{
\item{X}{a (preprocessed) n by px dataset}

\item{Y}{a (preprocessed) n by py dataset of responses}

\item{lv}{number of components}

\item{Q}{a py by py symmetric feature kernel representing similarities between features.}

\item{H}{a px by px symmetric sample kernel representing similarities between samples.}

\item{sumabsw}{a scalar contraint on the L1 norm of the weights.}

\item{deflateX}{(=FALSE default) whether to deflate X in the iterative estimation of components. Y is always deflated.}
}
\description{
Calculates a PLS model with a combined objective, maximising 1) covariance between X and Y, 2) description of feature to feature relationship, 3) Alternative kernel representation of sample-data and 4) with possible sparseness.
}
