% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structpls_pred.R
\name{structpls_pred}
\alias{structpls_pred}
\title{Structural PLS prediction}
\usage{
structpls_pred(model, newdata, lv = dim(B)[3])
}
\arguments{
\item{model}{the output from structpls()}

\item{newdata}{a (preprocessed) n by px dataset of predictors}

\item{lv}{number of components to use from the model (default = the maximum)}
}
\description{
Predicts Yhat for a structpls model
}
